+++ .eslintrc.yml
@@ -15,6 +15,7 @@ rules:
   prettier/prettier:
     - error
     - singleQuote: true
+      trailingComma: none
   camelcase: "off"
   '@typescript-eslint/naming-convention':
     - warn
@@ -58,6 +59,8 @@ rules:
   '@typescript-eslint/no-unsafe-member-access': off
   '@typescript-eslint/no-unsafe-argument': off
   '@typescript-eslint/no-unsafe-call': off
+  '@typescript-eslint/no-explicit-any': off
+  '@typescript-eslint/no-unused-vars': off
   'no-inner-declarations': off
   'no-global-assign': off
   'prefer-const': warn
