+++ README.md
@@ -4,7 +4,7 @@
 
 # Cubism Web Framework
 
-This is a framework for using models output by Live2D Cubism 4 Editor in applications.
+This is a framework for using models output by Live2D Cubism Editor in applications.
 
 It provides various functions for displaying and manipulating the model.
 It is used in conjunction with the Live2D Cubism Core library to load the model.
@@ -17,17 +17,26 @@ You can use it as a JavaScript library that can be used in the browser by buildi
 Please check the [license](LICENSE.md) before using this SDK.
 
 
+## Compatibility with Cubism 5 new features and previous Cubism SDK versions
+
+This SDK is compatible with Cubism 5.
+
+For SDK compatibility with new features in Cubism 5 Editor, please refer to [here](https://docs.live2d.com/en/cubism-sdk-manual/cubism-5-new-functions/).
+
+For compatibility with previous versions of Cubism SDK, please refer to [here](https://docs.live2d.com/en/cubism-sdk-manual/compatibility-with-cubism-5/).
+
+
 ## Development environment
 
 ### Node.js
 
-* 20.1.0
-* 18.16.0
-* 16.20.0
+* 20.5.1
+* 18.17.1
+* 16.20.2
 
 ### TypeScript
 
-5.0.4
+5.1.6
 
 
 ## Development environment construction
