+++ src/motion/cubismexpressionmotion.ts
@@ -31,6 +31,9 @@ const DefaultFadeTime = 1.0;
  * 表情のモーションクラス。
  */
 export class CubismExpressionMotion extends ACubismMotion {
+  static readonly DefaultAdditiveValue = 0.0; // 加算適用の初期値
+  static readonly DefaultMultiplyValue = 1.0; // 乗算適用の初期値
+
   /**
    * インスタンスを作成する。
    * @param buffer expファイルが読み込まれているバッファ
@@ -63,7 +66,7 @@ export class CubismExpressionMotion extends ACubismMotion {
       const parameter: ExpressionParameter = this._parameters.at(i);
 
       switch (parameter.blendType) {
-        case ExpressionBlendType.ExpressionBlendType_Add: {
+        case ExpressionBlendType.Additive: {
           model.addParameterValueById(
             parameter.parameterId,
             parameter.value,
@@ -71,7 +74,7 @@ export class CubismExpressionMotion extends ACubismMotion {
           );
           break;
         }
-        case ExpressionBlendType.ExpressionBlendType_Multiply: {
+        case ExpressionBlendType.Multiply: {
           model.multiplyParameterValueById(
             parameter.parameterId,
             parameter.value,
@@ -79,7 +82,7 @@ export class CubismExpressionMotion extends ACubismMotion {
           );
           break;
         }
-        case ExpressionBlendType.ExpressionBlendType_Overwrite: {
+        case ExpressionBlendType.Overwrite: {
           model.setParameterValueById(
             parameter.parameterId,
             parameter.value,
@@ -92,6 +95,163 @@ export class CubismExpressionMotion extends ACubismMotion {
           break;
       }
     }
+  }
+
+  /**
+   * @brief 表情によるモデルのパラメータの計算
+   *
+   * モデルの表情に関するパラメータを計算する。
+   *
+   * @param[in]   model                        対象のモデル
+   * @param[in]   userTimeSeconds              デルタ時間の積算値[秒]
+   * @param[in]   motionQueueEntry             CubismMotionQueueManagerで管理されているモーション
+   * @param[in]   expressionParameterValues    モデルに適用する各パラメータの値
+   * @param[in]   expressionIndex              表情のインデックス
+   */
+  public calculateExpressionParameters(
+    model: CubismModel,
+    userTimeSeconds: number,
+    motionQueueEntry: CubismMotionQueueEntry,
+    expressionParameterValues: csmVector<ExpressionParameterValue>,
+    expressionIndex: number
+  ) {
+    if (!motionQueueEntry.isAvailable()) {
+      return;
+    }
+
+    if (!motionQueueEntry.isStarted()) {
+      motionQueueEntry.setIsStarted(true);
+      motionQueueEntry.setStartTime(userTimeSeconds - this._offsetSeconds); // モーションの開始時刻を記録
+      motionQueueEntry.setFadeInStartTime(userTimeSeconds); // フェードインの開始時刻
+
+      const duration = this.getDuration();
+
+      if (motionQueueEntry.getEndTime() < 0.0) {
+        // 開始していないうちに終了設定している場合がある
+        motionQueueEntry.setEndTime(
+          duration <= 0.0 ? -1 : motionQueueEntry.getStartTime() + duration
+        );
+        // duration == -1 の場合はループする
+      }
+    }
+
+    this._fadeWeight = this.updateFadeWeight(motionQueueEntry, userTimeSeconds);
+
+    // モデルに適用する値を計算
+    for (let i = 0; i < expressionParameterValues.getSize(); ++i) {
+      const expressionParameterValue = expressionParameterValues.at(i);
+
+      if (expressionParameterValue.parameterId == null) {
+        continue;
+      }
+
+      const currentParameterValue = (expressionParameterValue.overwriteValue =
+        model.getParameterValueById(expressionParameterValue.parameterId));
+
+      const expressionParameters = this.getExpressionParameters();
+      let parameterIndex = -1;
+      for (let j = 0; j < expressionParameters.getSize(); ++j) {
+        if (
+          expressionParameterValue.parameterId !=
+          expressionParameters.at(j).parameterId
+        ) {
+          continue;
+        }
+
+        parameterIndex = j;
+
+        break;
+      }
+
+      // 再生中のExpressionが参照していないパラメータは初期値を適用
+      if (parameterIndex < 0) {
+        if (expressionIndex == 0) {
+          expressionParameterValue.additiveValue =
+            CubismExpressionMotion.DefaultAdditiveValue;
+          expressionParameterValue.multiplyValue =
+            CubismExpressionMotion.DefaultMultiplyValue;
+          expressionParameterValue.overwriteValue = currentParameterValue;
+        } else {
+          expressionParameterValue.additiveValue = this.calculateValue(
+            expressionParameterValue.additiveValue,
+            CubismExpressionMotion.DefaultAdditiveValue
+          );
+          expressionParameterValue.multiplyValue = this.calculateValue(
+            expressionParameterValue.multiplyValue,
+            CubismExpressionMotion.DefaultMultiplyValue
+          );
+          expressionParameterValue.overwriteValue = this.calculateValue(
+            expressionParameterValue.overwriteValue,
+            currentParameterValue
+          );
+        }
+        continue;
+      }
+
+      // 値を計算
+      const value = expressionParameters.at(parameterIndex).value;
+      let newAdditiveValue, newMultiplyValue, newOverwriteValue;
+      switch (expressionParameters.at(parameterIndex).blendType) {
+        case ExpressionBlendType.Additive:
+          newAdditiveValue = value;
+          newMultiplyValue = CubismExpressionMotion.DefaultMultiplyValue;
+          newOverwriteValue = currentParameterValue;
+          break;
+
+        case ExpressionBlendType.Multiply:
+          newAdditiveValue = CubismExpressionMotion.DefaultAdditiveValue;
+          newMultiplyValue = value;
+          newOverwriteValue = currentParameterValue;
+          break;
+
+        case ExpressionBlendType.Overwrite:
+          newAdditiveValue = CubismExpressionMotion.DefaultAdditiveValue;
+          newMultiplyValue = CubismExpressionMotion.DefaultMultiplyValue;
+          newOverwriteValue = value;
+          break;
+
+        default:
+          return;
+      }
+
+      if (expressionIndex == 0) {
+        expressionParameterValue.additiveValue = newAdditiveValue;
+        expressionParameterValue.multiplyValue = newMultiplyValue;
+        expressionParameterValue.overwriteValue = newOverwriteValue;
+      } else {
+        expressionParameterValue.additiveValue =
+          expressionParameterValue.additiveValue * (1.0 - this._fadeWeight) +
+          newAdditiveValue * this._fadeWeight;
+        expressionParameterValue.multiplyValue =
+          expressionParameterValue.multiplyValue * (1.0 - this._fadeWeight) +
+          newMultiplyValue * this._fadeWeight;
+        expressionParameterValue.overwriteValue =
+          expressionParameterValue.overwriteValue * (1.0 - this._fadeWeight) +
+          newOverwriteValue * this._fadeWeight;
+      }
+    }
+  }
+
+  /**
+   * @brief 表情が参照しているパラメータを取得
+   *
+   * 表情が参照しているパラメータを取得する
+   *
+   * @return 表情パラメータ
+   */
+  public getExpressionParameters() {
+    return this._parameters;
+  }
+
+  /**
+   * @brief 表情のフェードの値を取得
+   *
+   * 現在の表情のフェードのウェイト値を取得する
+   *
+   * @returns 表情のフェードのウェイト値
+   */
+  public getFadeWeight() {
+    return this._fadeWeight;
   }
 
   protected parse(buffer: ArrayBuffer, size: number) {
@@ -130,20 +290,20 @@ export class CubismExpressionMotion extends ACubismMotion {
         param.getValueByString(ExpressionKeyBlend).isNull() ||
         param.getValueByString(ExpressionKeyBlend).getString() == BlendValueAdd
       ) {
-        blendType = ExpressionBlendType.ExpressionBlendType_Add;
+        blendType = ExpressionBlendType.Additive;
       } else if (
         param.getValueByString(ExpressionKeyBlend).getString() ==
         BlendValueMultiply
       ) {
-        blendType = ExpressionBlendType.ExpressionBlendType_Multiply;
+        blendType = ExpressionBlendType.Multiply;
       } else if (
         param.getValueByString(ExpressionKeyBlend).getString() ==
         BlendValueOverwrite
       ) {
-        blendType = ExpressionBlendType.ExpressionBlendType_Overwrite;
+        blendType = ExpressionBlendType.Overwrite;
       } else {
         // その他 仕様にない値を設定した時は加算モードにすることで復旧
-        blendType = ExpressionBlendType.ExpressionBlendType_Add;
+        blendType = ExpressionBlendType.Additive;
       }
 
       // 設定オブジェクトを作成してリストに追加する
@@ -160,24 +320,39 @@ export class CubismExpressionMotion extends ACubismMotion {
   }
 
   /**
+   * @brief ブレンド計算
+   *
+   * 入力された値でブレンド計算をする。
+   *
+   * @param source 現在の値
+   * @param destination 適用する値
+   * @param weight ウェイト
+   * @returns 計算結果
+   */
+  public calculateValue(source: number, destination: number): number {
+    return source * (1.0 - this._fadeWeight) + destination * this._fadeWeight;
+  }
+
+  /**
    * コンストラクタ
    */
   protected constructor() {
     super();
-
     this._parameters = new csmVector<ExpressionParameter>();
+    this._fadeWeight = 0.0;
   }
 
-  _parameters: csmVector<ExpressionParameter>; // 表情のパラメータ情報リスト
+  private _parameters: csmVector<ExpressionParameter>; // 表情のパラメータ情報リスト
+  private _fadeWeight: number; // 表情の現在のウェイト
 }
 
 /**
  * 表情パラメータ値の計算方式
  */
 export enum ExpressionBlendType {
-  ExpressionBlendType_Add = 0, // 加算
-  ExpressionBlendType_Multiply = 1, // 乗算
-  ExpressionBlendType_Overwrite = 2, // 上書き
+  Additive = 0, // 加算
+  Multiply = 1, // 乗算
+  Overwrite = 2 // 上書き
 }
 
 /**
@@ -191,6 +366,8 @@ export class ExpressionParameter {
 
 // Namespace definition for compatibility.
 import * as $ from './cubismexpressionmotion';
+import { ExpressionParameterValue } from './cubismexpressionmotionmanager';
+import { CubismDefaultParameterId } from '../cubismdefaultparameterid';
 // eslint-disable-next-line @typescript-eslint/no-namespace
 export namespace Live2DCubismFramework {
   export const CubismExpressionMotion = $.CubismExpressionMotion;
