+++ src/motion/cubismmotioninternal.ts
@@ -17,7 +17,7 @@ import { csmVector } from '../type/csmvector';
 export enum CubismMotionCurveTarget {
   CubismMotionCurveTarget_Model, // モデルに対して
   CubismMotionCurveTarget_Parameter, // パラメータに対して
-  CubismMotionCurveTarget_PartOpacity, // パーツの不透明度に対して
+  CubismMotionCurveTarget_PartOpacity // パーツの不透明度に対して
 }
 
 /**
@@ -29,7 +29,7 @@ export enum CubismMotionSegmentType {
   CubismMotionSegmentType_Linear = 0, // リニア
   CubismMotionSegmentType_Bezier = 1, // ベジェ曲線
   CubismMotionSegmentType_Stepped = 2, // ステップ
-  CubismMotionSegmentType_InverseStepped = 3, // インバースステップ
+  CubismMotionSegmentType_InverseStepped = 3 // インバースステップ
 }
 
 /**
