+++ src/physics/cubismphysicsinternal.ts
@@ -13,7 +13,7 @@ import { csmVector } from '../type/csmvector';
  * 物理演算の適用先の種類
  */
 export enum CubismPhysicsTargetType {
-  CubismPhysicsTargetType_Parameter, // パラメータに対して適用
+  CubismPhysicsTargetType_Parameter // パラメータに対して適用
 }
 
 /**
@@ -22,7 +22,7 @@ export enum CubismPhysicsTargetType {
 export enum CubismPhysicsSource {
   CubismPhysicsSource_X, // X軸の位置から
   CubismPhysicsSource_Y, // Y軸の位置から
-  CubismPhysicsSource_Angle, // 角度から
+  CubismPhysicsSource_Angle // 角度から
 }
 
 /**
